Moved tests to root dir
